Доработка согласно комментариям ревьюера. Добавлены фикстуры для создания и удаления курьера. Тесты разнесены по отдельным файлам
