+++ README.md
@@ -8,19 +8,22 @@
 
 ## Результаты тестов
 
-Результаты тестов сохраняются в формате Allure и находятся в директории `tests/allure_results/`. Для просмотра результатов тестов используйте Allure Report
+Результаты тестов сохраняются в формате Allure и находятся в директории `allure_results/`. Для просмотра результатов тестов используйте Allure Report
 
 ## Структура проекта
 
-- `data.py`: Содержит данные, используемые в тестах
-- `helpers.py`: Содержит вспомогательные функции для тестов
+- `data.py`: Содержит базовый URL запроса и тело запроса для создания заказа
+- `helpers.py`: Содержит вспомогательные функции для тестов, такие как генерация случайных строк, создание тела запроса для курьера и удаление созданных данных
 - `tests/`: Директория с тестами
-  - `tests.py`: Файл с тест-кейсами
-  - `allure_results/`: Директория с результатами тестов в формате Allure
+  - `test_create_courier.py`: Тесты для создания курьеров, проверка статуса ответа, тела ответа и уникальности логинов
+  - `test_create_order.py`: Тесты для создания заказов
+  - `test_login.py`: Тесты для авторизации курьеров
+  - `test_order_list.py`: Тесты для получения списка заказов
+- `allure_results/`: Директория с результатами тестов в формате Allure
 
 ## Зависимости
 
-Для установки зависимостей используйте файл `tests/requirements.txt`:
+Для установки зависимостей используйте файл `requirements.txt`:
 ```
 allure-pytest==2.14.3
 allure-python-commons==2.14.3
