+++ helpers.py
@@ -4,74 +4,11 @@ import string
 import data
 
 class Helpers:
-    # метод регистрации нового курьера возвращает список из логина и пароля
-    # если регистрация не удалась, возвращает пустой список
-    def register_new_courier_and_return_login_password():
-    # метод генерирует строку, состоящую только из букв нижнего регистра, в качестве параметра передаём длину строки
-        def generate_random_string(length):
-            letters = string.ascii_lowercase
-            random_string = ''.join(random.choice(letters) for i in range(length))
-            return random_string
-
-        # создаём список, чтобы метод мог его вернуть
-        login_pass = []
-
-        # генерируем логин, пароль и имя курьера
-        login = generate_random_string(10)
-        password = generate_random_string(10)
-        first_name = generate_random_string(10)
-
-        # собираем тело запроса
-        payload = {
-            "login": login,
-            "password": password,
-            "firstName": first_name
-        }
-
-        # отправляем запрос на регистрацию курьера и сохраняем ответ в переменную response
-        response = requests.post(data.URL+'/courier', data=payload)
-
-        # если регистрация прошла успешно (код ответа 201), добавляем в список логин и пароль курьера
-        if response.status_code == 201:
-            login_pass.append(login)
-            login_pass.append(password)
-            login_pass.append(first_name)
-
-        # возвращаем список
-        return login_pass
-
-
     # Метод для генерации строк заданной длины length
     def generate_random_string(length):
         letters = string.ascii_lowercase
         random_string = ''.join(random.choice(letters) for i in range(length))
         return random_string
-
-
-    # метод регистрирует нового курьера и возвращает тело ответа
-    def register_new_courier():
-        # метод генерирует строку, состоящую только из букв нижнего регистра, в качестве параметра передаём длину строки
-        def generate_random_string(length):
-            letters = string.ascii_lowercase
-            random_string = ''.join(random.choice(letters) for i in range(length))
-            return random_string
-
-        # генерируем логин, пароль и имя курьера
-        login = generate_random_string(10)
-        password = generate_random_string(10)
-        first_name = generate_random_string(10)
-
-        # собираем тело запроса
-        payload = {
-            "login": login,
-            "password": password,
-            "firstName": first_name
-        }
-
-        # отправляем запрос на регистрацию курьера и сохраняем ответ в переменную response
-        response = requests.post(data.URL+'/courier', data=payload)
-
-        return response
 
 
     # Функция создания тела запроса для курьера
@@ -81,3 +18,9 @@ class Helpers:
             "password": password
         }
         return payload
+    
+
+    # Функция для удаления созданных данных
+    def delete_courier(id):
+        response = requests.delete(f"{data.URL}/courier/{id}")
+        return response
